+++ fittrackee/tests/application/test_config.py
@@ -41,15 +41,12 @@ class TestTestingConfig:
         app.config.from_object('fittrackee.config.TestingConfig')
 
         assert app.config['SQLALCHEMY_DATABASE_URI'] == os.environ.get(
-            'DATABASE_TEST_URL'
+            'DATABASE_TEST_URL', ''
+        ) + (
+            f"_{os.getenv('PYTEST_XDIST_WORKER')}"
+            if os.getenv('PYTEST_XDIST_WORKER')
+            else ''
         )
-
-    def test_it_does_not_preserve_context_on_exception(
-        self, app: Flask
-    ) -> None:
-        app.config.from_object('fittrackee.config.TestingConfig')
-
-        assert not app.config['PRESERVE_CONTEXT_ON_EXCEPTION']
 
 
 class TestProductionConfig:
@@ -71,10 +68,3 @@ class TestProductionConfig:
         assert app.config['SQLALCHEMY_DATABASE_URI'] == os.environ.get(
             'DATABASE_TEST_URL'
         )
-
-    def test_it_does_not_preserve_context_on_exception(
-        self, app: Flask
-    ) -> None:
-        app.config.from_object('fittrackee.config.ProductionConfig')
-
-        assert not app.config['PRESERVE_CONTEXT_ON_EXCEPTION']