+++ fittrackee/tests/conftest.py
@@ -8,6 +8,7 @@ os.environ['APP_SETTINGS'] = 'fittrackee.config.TestingConfig'
 # to avoid resetting dev database during tests
 os.environ['DATABASE_URL'] = os.environ['DATABASE_TEST_URL']
 TEMP_FOLDER = '/tmp/FitTrackee'
+os.makedirs(TEMP_FOLDER, exist_ok=True)
 os.environ['UPLOAD_FOLDER'] = TEMP_FOLDER
 os.environ['APP_LOG'] = TEMP_FOLDER + '/fittrackee.log'
 os.environ['AUTHLIB_INSECURE_TRANSPORT'] = '1'