+++ fittrackee/tests/workouts/test_workouts_api_1_post.py
@@ -3,7 +3,7 @@ import os
 import re
 from datetime import datetime
 from io import BytesIO
-from typing import Dict, Optional
+from typing import Any, Dict, Optional
 from unittest.mock import Mock
 
 import pytest
@@ -446,12 +446,37 @@ class TestPostWorkoutWithGpx(ApiTestCaseMixin, CallArgsMixin):
         assert len(data['data']['workouts']) == 1
         assert_workout_data_with_gpx(data)
 
+    def test_it_returns_400_when_quotes_are_not_escaped_in_notes(
+        self,
+        app: Flask,
+        user_1: User,
+        sport_1_cycling: Sport,
+        gpx_file: str,
+    ) -> None:
+        client, auth_token = self.get_test_client_and_auth_token(
+            app, user_1.email
+        )
+
+        response = client.post(
+            '/api/workouts',
+            data=dict(
+                file=(BytesIO(str.encode(gpx_file)), 'example.gpx'),
+                data='{{"sport_id": 1, "notes": "test "workout""}}',
+            ),
+            headers=dict(
+                content_type='multipart/form-data',
+                Authorization=f'Bearer {auth_token}',
+            ),
+        )
+
+        self.assert_400(response)
+
     @pytest.mark.parametrize(
         'input_description,input_notes',
         [
             ('empty notes', ''),
             ('short notes', 'test workout'),
-            ('notes with special characters', 'test \nworkout'),
+            ('notes with special characters', "test \n'workout'"),
         ],
     )
     def test_it_adds_a_workout_with_gpx_notes(
@@ -861,8 +886,21 @@ class TestPostWorkoutWithoutGpx(ApiTestCaseMixin):
         assert len(data['data']['workouts']) == 1
         assert_workout_data_wo_gpx(data)
 
-    def test_it_returns_400_if_workout_date_is_missing(
-        self, app: Flask, user_1: User, sport_1_cycling: Sport
+    @pytest.mark.parametrize(
+        'input_ascent, input_descent',
+        [
+            (100, 150),
+            (0, 150),
+            (100, 0),
+        ],
+    )
+    def test_it_adds_workout_with_ascent_and_descent_when_provided(
+        self,
+        app: Flask,
+        user_1: User,
+        sport_1_cycling: Sport,
+        input_ascent: int,
+        input_descent: int,
     ) -> None:
         client, auth_token = self.get_test_client_and_auth_token(
             app, user_1.email
@@ -871,13 +909,157 @@ class TestPostWorkoutWithoutGpx(ApiTestCaseMixin):
         response = client.post(
             '/api/workouts/no_gpx',
             content_type='application/json',
-            data=json.dumps(dict(sport_id=1, duration=3600, distance=10)),
+            data=json.dumps(
+                dict(
+                    sport_id=1,
+                    duration=3600,
+                    workout_date='2018-05-15 14:05',
+                    distance=10,
+                    ascent=input_ascent,
+                    descent=input_descent,
+                )
+            ),
+            headers=dict(Authorization=f'Bearer {auth_token}'),
+        )
+
+        data = json.loads(response.data.decode())
+        assert response.status_code == 201
+        assert 'created' in data['status']
+        assert len(data['data']['workouts']) == 1
+        assert data['data']['workouts'][0]['ascent'] == input_ascent
+        assert data['data']['workouts'][0]['descent'] == input_descent
+
+    @pytest.mark.parametrize(
+        'description,input_data',
+        [
+            (
+                "'sport_id' is missing",
+                {
+                    "duration": 3600,
+                    "workout_date": '2018-05-15 14:05',
+                    "distance": 10,
+                },
+            ),
+            (
+                "'duration' is missing",
+                {
+                    "sport_id": 1,
+                    "workout_date": '2018-05-15 14:05',
+                    "distance": 10,
+                },
+            ),
+            (
+                "'workout_date' is missing",
+                {"sport_id": 1, "duration": 3600, "distance": 10},
+            ),
+            (
+                "'distance' is missing",
+                {
+                    "sport_id": 1,
+                    "duration": 3600,
+                    "workout_date": '2018-05-15 14:05',
+                },
+            ),
+        ],
+    )
+    def test_it_returns_400_if_key_is_missing(
+        self,
+        app: Flask,
+        user_1: User,
+        sport_1_cycling: Sport,
+        description: str,
+        input_data: Dict,
+    ) -> None:
+        client, auth_token = self.get_test_client_and_auth_token(
+            app, user_1.email
+        )
+
+        response = client.post(
+            '/api/workouts/no_gpx',
+            content_type='application/json',
+            data=json.dumps(input_data),
             headers=dict(Authorization=f'Bearer {auth_token}'),
         )
 
         self.assert_400(response)
 
-    def test_it_returns_500_if_workout_format_is_invalid(
+    @pytest.mark.parametrize(
+        'description,input_data',
+        [
+            ("only ascent", {"ascent": 100}),
+            ("only descent", {"descent": 150}),
+        ],
+    )
+    def test_it_returns_400_when_ascent_or_descent_are_missing(
+        self,
+        app: Flask,
+        user_1: User,
+        sport_1_cycling: Sport,
+        description: str,
+        input_data: Dict,
+    ) -> None:
+        client, auth_token = self.get_test_client_and_auth_token(
+            app, user_1.email
+        )
+
+        response = client.post(
+            '/api/workouts/no_gpx',
+            content_type='application/json',
+            data=json.dumps(
+                {
+                    'sport_id': 1,
+                    'duration': 3600,
+                    'workout_date': '2018-05-15 14:05',
+                    'distance': 10,
+                    **input_data,
+                }
+            ),
+            headers=dict(Authorization=f'Bearer {auth_token}'),
+        )
+
+        self.assert_400(response)
+
+    @pytest.mark.parametrize(
+        'description,input_data',
+        [
+            ("ascent is below 0", {"ascent": -100, "descent": 100}),
+            ("descent is below 0", {"ascent": 150, "descent": -100}),
+            ("ascent is None", {"ascent": None, "descent": 100}),
+            ("descent is None", {"ascent": 150, "descent": None}),
+            ("ascent is invalid", {"ascent": "a", "descent": 100}),
+            ("descent is invalid", {"ascent": 150, "descent": "b"}),
+        ],
+    )
+    def test_it_returns_400_when_ascent_or_descent_are_invalid(
+        self,
+        app: Flask,
+        user_1: User,
+        sport_1_cycling: Sport,
+        description: str,
+        input_data: Dict,
+    ) -> None:
+        client, auth_token = self.get_test_client_and_auth_token(
+            app, user_1.email
+        )
+
+        response = client.post(
+            '/api/workouts/no_gpx',
+            content_type='application/json',
+            data=json.dumps(
+                {
+                    'sport_id': 1,
+                    'duration': 3600,
+                    'workout_date': '2018-05-15 14:05',
+                    'distance': 10,
+                    **input_data,
+                }
+            ),
+            headers=dict(Authorization=f'Bearer {auth_token}'),
+        )
+
+        self.assert_400(response)
+
+    def test_it_returns_500_if_workout_date_format_is_invalid(
         self, app: Flask, user_1: User, sport_1_cycling: Sport
     ) -> None:
         client, auth_token = self.get_test_client_and_auth_token(
@@ -900,12 +1082,13 @@ class TestPostWorkoutWithoutGpx(ApiTestCaseMixin):
 
         self.assert_500(response, 'Error during workout save.', status='fail')
 
-    def test_it_adds_workout_with_zero_value(
+    @pytest.mark.parametrize('input_distance', [0, '', None])
+    def test_it_returns_400_when_distance_is_invalid(
         self,
         app: Flask,
         user_1: User,
         sport_1_cycling: Sport,
-        sport_2_running: Sport,
+        input_distance: Any,
     ) -> None:
         client, auth_token = self.get_test_client_and_auth_token(
             app, user_1.email
@@ -917,41 +1100,43 @@ class TestPostWorkoutWithoutGpx(ApiTestCaseMixin):
             data=json.dumps(
                 dict(
                     sport_id=1,
-                    duration=0,
-                    workout_date='2018-05-14 14:05',
-                    distance=0,
-                    title='Workout test',
+                    duration=3600,
+                    workout_date='2018-05-15 14:05',
+                    distance=input_distance,
                 )
             ),
             headers=dict(Authorization=f'Bearer {auth_token}'),
         )
 
-        data = json.loads(response.data.decode())
-        assert response.status_code == 201
-        assert 'created' in data['status']
-        assert len(data['data']['workouts']) == 1
-        assert 'creation_date' in data['data']['workouts'][0]
-        assert (
-            data['data']['workouts'][0]['workout_date']
-            == 'Mon, 14 May 2018 14:05:00 GMT'
-        )
-        assert data['data']['workouts'][0]['user'] == 'test'
-        assert data['data']['workouts'][0]['sport_id'] == 1
-        assert data['data']['workouts'][0]['duration'] is None
-        assert data['data']['workouts'][0]['title'] == 'Workout test'
-        assert data['data']['workouts'][0]['ascent'] is None
-        assert data['data']['workouts'][0]['ave_speed'] is None
-        assert data['data']['workouts'][0]['descent'] is None
-        assert data['data']['workouts'][0]['distance'] is None
-        assert data['data']['workouts'][0]['max_alt'] is None
-        assert data['data']['workouts'][0]['max_speed'] is None
-        assert data['data']['workouts'][0]['min_alt'] is None
-        assert data['data']['workouts'][0]['moving'] is None
-        assert data['data']['workouts'][0]['pauses'] is None
-        assert data['data']['workouts'][0]['with_gpx'] is False
+        self.assert_400(response)
 
-        assert len(data['data']['workouts'][0]['segments']) == 0
-        assert len(data['data']['workouts'][0]['records']) == 0
+    @pytest.mark.parametrize('input_duration', [0, '', None])
+    def test_it_returns_400_when_duration_is_invalid(
+        self,
+        app: Flask,
+        user_1: User,
+        sport_1_cycling: Sport,
+        input_duration: Any,
+    ) -> None:
+        client, auth_token = self.get_test_client_and_auth_token(
+            app, user_1.email
+        )
+
+        response = client.post(
+            '/api/workouts/no_gpx',
+            content_type='application/json',
+            data=json.dumps(
+                dict(
+                    sport_id=1,
+                    duration=input_duration,
+                    workout_date='2018-05-15 14:05',
+                    distance=10,
+                )
+            ),
+            headers=dict(Authorization=f'Bearer {auth_token}'),
+        )
+
+        self.assert_400(response)
 
     @pytest.mark.parametrize(
         'client_scope, can_access',
@@ -1112,7 +1297,7 @@ class TestPostWorkoutWithZipArchive(ApiTestCaseMixin):
             data = self.assert_500(response, 'error during gpx processing')
             assert 'data' not in data
 
-    def test_it_imports_only_max_number_of_files(
+    def test_it_returns_400_when_files_in_archive_exceed_limit(
         self,
         app_with_max_workouts: Flask,
         user_1: User,
@@ -1127,7 +1312,7 @@ class TestPostWorkoutWithZipArchive(ApiTestCaseMixin):
                 app_with_max_workouts, user_1.email
             )
 
-            client.post(
+            response = client.post(
                 '/api/workouts',
                 data=dict(
                     file=(zip_file, 'gpx_test.zip'), data='{"sport_id": 1}'
@@ -1138,12 +1323,11 @@ class TestPostWorkoutWithZipArchive(ApiTestCaseMixin):
                 ),
             )
 
-            response = client.get(
-                '/api/workouts',
-                headers=dict(Authorization=f'Bearer {auth_token}'),
+            self.assert_400(
+                response,
+                'the number of files in the archive exceeds the limit',
+                'fail',
             )
-            data = json.loads(response.data.decode())
-            assert len(data['data']['workouts']) == 2
 
     def test_it_returns_error_if_archive_size_exceeds_limit(
         self,
@@ -1175,6 +1359,40 @@ class TestPostWorkoutWithZipArchive(ApiTestCaseMixin):
                 response,
                 'Error during workout upload, '
                 'file size (2.5KB) exceeds 1.0KB.',
+            )
+            assert 'data' not in data
+
+    def test_it_returns_error_if_a_file_from_archive_size_exceeds_limit(
+        self,
+        app_with_max_file_size: Flask,
+        user_1: User,
+        sport_1_cycling: Sport,
+    ) -> None:
+        file_path = os.path.join(
+            app_with_max_file_size.root_path, 'tests/files/gpx_test.zip'
+        )
+        # 'gpx_test.zip' contains 3 gpx files (same data) and 1 non-gpx file
+        with open(file_path, 'rb') as zip_file:
+            client, auth_token = self.get_test_client_and_auth_token(
+                app_with_max_file_size, user_1.email
+            )
+
+            response = client.post(
+                '/api/workouts',
+                data=dict(
+                    file=(zip_file, 'gpx_test.zip'), data='{"sport_id": 1}'
+                ),
+                headers=dict(
+                    content_type='multipart/form-data',
+                    Authorization=f'Bearer {auth_token}',
+                ),
+            )
+
+            data = self.assert_400(
+                response,
+                'at least one file in zip archive exceeds size limit, '
+                'please check the archive',
+                'fail',
             )
             assert 'data' not in data