+++ fittrackee/tests/workouts/test_workouts_model.py
@@ -75,6 +75,48 @@ class TestWorkoutModel:
         assert serialized_workout['with_gpx'] is False
         assert str(serialized_workout['workout_date']) == '2018-01-01 00:00:00'
 
+    def test_serialize_for_workout_without_gpx_and_with_ascent_and_descent(
+        self,
+        app: Flask,
+        sport_1_cycling: Sport,
+        user_1: User,
+        workout_cycling_user_1: Workout,
+    ) -> None:
+        workout = workout_cycling_user_1
+        workout.ascent = 0
+        workout.descent = 10
+
+        serialized_workout = workout.serialize()
+        assert serialized_workout['ascent'] == workout.ascent
+        assert serialized_workout['ave_speed'] == float(workout.ave_speed)
+        assert serialized_workout['bounds'] == []
+        assert 'creation_date' in serialized_workout
+        assert serialized_workout['descent'] == workout.descent
+        assert serialized_workout['distance'] == float(workout.distance)
+        assert serialized_workout['duration'] == str(workout.duration)
+        assert serialized_workout['id'] == workout.short_id
+        assert serialized_workout['map'] is None
+        assert serialized_workout['max_alt'] is None
+        assert serialized_workout['max_speed'] == float(workout.max_speed)
+        assert serialized_workout['min_alt'] is None
+        assert serialized_workout['modification_date'] is not None
+        assert serialized_workout['moving'] == str(workout.moving)
+        assert serialized_workout['next_workout'] is None
+        assert serialized_workout['notes'] is None
+        assert serialized_workout['pauses'] is None
+        assert serialized_workout['previous_workout'] is None
+        assert serialized_workout['records'] == [
+            record.serialize() for record in workout.records
+        ]
+        assert serialized_workout['segments'] == []
+        assert serialized_workout['sport_id'] == workout.sport_id
+        assert serialized_workout['title'] == workout.title
+        assert serialized_workout['user'] == workout.user.username
+        assert serialized_workout['weather_end'] is None
+        assert serialized_workout['weather_start'] is None
+        assert serialized_workout['with_gpx'] is False
+        assert str(serialized_workout['workout_date']) == '2018-01-01 00:00:00'
+
     def test_serialize_for_workout_with_gpx(
         self,
         app: Flask,