+++ fittrackee_client/src/components/User/ProfileEdition/UserPreferencesEdition.vue
@@ -242,5 +242,10 @@
         }
       }
     }
+
+    #language,
+    #date_format {
+      padding: $default-padding * 0.5;
+    }
   }
 </style>