+++ fittrackee_client/src/utils/dates.ts
@@ -97,7 +97,9 @@ const dateStringFormats: Record<string, string> = {
   de: 'do MMM yyyy',
   en: 'MMM. do, yyyy',
   fr: 'd MMM yyyy',
+  it: 'd MMM yyyy',
   // nb: 'do MMM yyyy',
+  nl: 'd MMM yyyy',
 }
 
 export const getDateFormat = (dateFormat: string, language: string): string => {