+++ fittrackee_client/src/utils/locales.ts
@@ -1,6 +1,6 @@
 /* eslint-disable import/no-duplicates */
 import { Locale } from 'date-fns'
-import { de, enUS, fr } from 'date-fns/locale'
+import { de, enUS, fr, it, nl } from 'date-fns/locale'
 
 import createI18n from '@/i18n'
 
@@ -8,14 +8,18 @@ export const localeFromLanguage: Record<string, Locale> = {
   de: de,
   en: enUS,
   fr: fr,
+  it: it,
   // nb: nb, // disabled for now
+  nl: nl,
 }
 
 export const languageLabels: Record<string, string> = {
   de: 'Deutsch',
   en: 'English',
   fr: 'Français',
+  it: 'Italiano',
   // nb: 'Norsk bokmål', // disabled for now
+  nl: 'Nederlands',
 }
 
 const { availableLocales } = createI18n.global